requirements.txt
selenium>=4.16.0
pytest>=7.4.4
webdriver-manager>=4.0.2
pytest-html>=4.1.1
pytest-xdist>=3.7.0
allure-pytest==2.13.5
allure-python-commons==2.13.5

pandas==2.3.0
openpyxl==3.1.5

